chore: migrate engine state-store modules to TypeScript (closes #373)
Convert the engine state-store layer to TypeScript as a type-only migration
with no change to public APIs or runtime behavior:

- shared_state_store, memory/memcached/redis_state_store, session_state,
  session_live_state, playhead_state (.js -> .ts via git mv).
- Add shared state-store type definitions (IStateStore, ISharedStateStoreOpts,
  ILeaderCache, SharedStates) in engine/state_store_types.ts.
- Keep CommonJS require/module.exports interop (erasable syntax) so the still
  -.js consumers and specs load the modules unchanged.
- Add a jasmine helper that resolves .js requires to .ts sources during the
  source-tree test run (build/test-time shim only).
